generics and navbar component
