+++ src/App.css
@@ -2,11 +2,13 @@
     background-color: #1C1917;
     color: white;
     min-height: 100vh;
-    padding: 0px 50px 40px 50px;
+    padding: 0px 50px 60px 50px;
 }
 
 nav {
-    padding: 30px 0px;
+    height: 116px;
+    display: flex;
+    align-items: center;
 }
 
 .flex {
@@ -15,6 +17,7 @@ nav {
     justify-content: space-between;
     margin: auto;
     flex-wrap: wrap;
+    width: 100%;
     max-width: 1250px;
     gap: 25px;
 }
