+++ src/components/Recommended.tsx
@@ -13,11 +13,14 @@ import "../styles/Recipes.css";
 // Hooks
 import useFetch from '../hooks/useFetch';
 
+// Types
+import TypeRecipe from "../types/TypeRecipe";
+
 //.env
 const apiUrl = import.meta.env.VITE_API_URL;
 
 function Recipes() {
-    const { recipes, loading, error } = useFetch(`${apiUrl}recipes/random?limit=10`);
+    const { recipes, loading, error } = useFetch<TypeRecipe[]>(`${apiUrl}recipes/random?limit=10`);
 
     const prevRef = useRef(null);
     const nextRef = useRef(null);
@@ -28,7 +31,7 @@ function Recipes() {
                 <p>Loading...</p>
             ) : error ? (
                 <p>{error}</p>
-            ) : recipes.length === 0 ? (
+            ) : recipes?.length === 0 ? (
                 <p>No recipes found!</p>
             ) : (
                 <div className="recipes-slider">
@@ -66,9 +69,9 @@ function Recipes() {
                         }}
                         className="recommend"
                     >
-                        {recipes.map((recipe) => (
+                        {recipes?.map((recipe) => (
                             <SwiperSlide key={recipe._id}>
-                                <Link to="/" style={{ color: 'white' }}>
+                                <Link to={`/recipe/${recipe._id}`} style={{ color: 'white' }}>
                                     <div className="recipeTop">
                                         <p>{recipe.cuisine}</p>
                                         <button className="mark">
