+++ src/styles/Recipes.css
@@ -1,5 +1,5 @@
 .Recipes {
-    padding: 60px 0px 60px 0px;
+    padding: 55px 0px;
 }
 
 .recipes-slider {
